Separate Codelet.serialise() in to two methods
The concerns of serialising a codelet into JSON and writing it to
file/stdout were combined in the `serialise()` method.

This commit reduces the scope of the Codelet.serialise() method
to just returning serialised JSON.

It creates a new method, Codelete.write(), to write JSON to file/stdout.

This makes it easier to test the JSON returned by serialise().

Caller code is also (IMO) more readable, e.g.
`codetree.write(self._args.output)`